+++ src/steps/build-project-user-relationships/index.ts
@@ -77,7 +77,7 @@ async function createUserIdMap(
   return userIdMap;
 }
 
-export default createStep((instance) => createGitlabClient(instance));
+export default createStep((instance) => createGitlabClient(instance.config));
 
 export function createProjectUserRelationship(
   project: Entity,